refactor: decompose bulk-add GroupReviewStep and PhotoGroupingGrid into focused modules
Extract AI streaming logic into useGroupReviewAi, drag-and-drop into usePhotoGridDrag, and
UI pieces (AiCorrectionWidget, AiRetryBand, GroupReviewForm, GroupReviewPhoto, BinStack,
StackedPhotoTile, AddMoreTile, SplitZone, DragGhost) into their own files. Shared photo
compression helpers consolidated into aiPhotoFormData.ts.
